Move warnings into 2and3
